pyproject.toml
[project]
name = "teste-tecnico"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.0",
    "langchain-qdrant>=1.1.0",
    "pypdf>=6.4.0",
    "qdrant-client>=1.16.0",
    "rich>=14.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.8.0",
    "langfuse>=2.0.0",
    "langfuse-langchain>=2.0.0",
    "langchain>=1.1.0",
]
ignore = [
  "T201",   # Checks for print statements, 
  "COM812", # Checks for the absence of trailing commas.
  "INP001", # Checks for packages that are missing an __init__.py file.
  "D",      # All pydocstyle (D)
  "ANN401", # Checks that function arguments are annotated with a more specific type than Any.
  "ERA001", # Checks for commented-out Python code.
  "A004",   # Shadowing Python Builtin
]
[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "T201",   # Checks for print statements, 
  "COM812", # Checks for the absence of trailing commas.
  "INP001", # Checks for packages that are missing an __init__.py file.
  "D",      # All pydocstyle (D)
  "ANN401", # Checks that function arguments are annotated with a more specific type than Any.
  "ERA001", # Checks for commented-out Python code.
  "A004",   # Shadowing Python Builtin
  "E402",   # Module level import not at top of file (necessário para sys.path setup)
  "I001",   # Import block is un-sorted or un-formatted
]
